Adds string indirection for resampling API enum

+++ samples/helpers.h
@@ -283,7 +283,7 @@ struct Surface {
         }
         initImage(hostPtr, n_elems);
         uint32_t *temp = (uint32_t *)hostPtr;
-        for (size_t i = 0; i < n_elems; i = i+2) {
+        for (int64_t i = 0; i < n_elems; i = i+2) {
             temp[i + 1] = 1u;
         }
             checkCudaErr(cudaMemcpy(devPtr, hostPtr, size_t(sizeof(hostPtr[0]) * n_elems), cudaMemcpyHostToDevice));
